// web/src/server/predict/types.ts
export type ProbabilityTriplet = {
  home: number;
  draw: number;
  away: number;
};

export type FixtureTeam = {
  canonical: string;
  name: string;
  shortName: string;
  crest: string;
  league: string;
};

export type FinancialFeatures = {
  wageGap: number;
  netSpendGap: number;
  valuationGap: number;
};

export type MarketFeatures = {
  marketEdge: number;
  volatility: number;
};

export type PerformanceFeatures = {
  attGap: number;
  defGap: number;
  xgGap: number;
  xgDefGap: number;
  pointsGap: number;
};

export type FixtureFeatureVector = PerformanceFeatures &
  FinancialFeatures &
  MarketFeatures & {
    homeForm: number;
    awayForm: number;
    [feature: string]: number;
  };

export type FixtureContext = {
  season: string;
  home: FixtureTeam;
  away: FixtureTeam;
};